+++ mongooseConnectionOptions.ts
@@ -2,16 +2,15 @@
  * mongoose接続オプション
  * @see http://mongoosejs.com/docs/api.html#index_Mongoose-connect
  * @see http://mongodb.github.io/node-mongodb-native/2.2/api/MongoClient.html
- * @ignore
  */
 const mongooseConnectionOptions = {
-    useMongoClient: true,
     autoReconnect: true,
-    keepAlive: 120000,
+    keepAlive: true,
     connectTimeoutMS: 30000,
     socketTimeoutMS: 0,
     reconnectTries: 30,
-    reconnectInterval: 1000
+    reconnectInterval: 1000,
+    useNewUrlParser: true
 };
 
 export default mongooseConnectionOptions;